% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SolveRationalMatrixEquaton.R
\name{LQdecompose}
\alias{LQdecompose}
\title{LQ Decomposition}
\usage{
LQdecompose(mat)
}
\arguments{
\item{mat}{a matrix of real numbers}
}
\value{
list containing L and Q matrices
}
\description{
Decompose a matrix into a Lower triangular matrix L and an orthogonal matrix Q
}
\examples{
LQdecompose(rbind(c(2,-2,18),c(2,1,0),c(1,2,0)))
}
