% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SolveRationalMatrixEquaton.R
\name{QRdecompose}
\alias{QRdecompose}
\title{QR Decomposition}
\usage{
QRdecompose(mat)
}
\arguments{
\item{mat}{a matrix of real numbers}
}
\value{
list containing Q and R matrices
}
\description{
Decompose a matrix into an orthogonal matrix Q and Upper triangular matrix R
}
\examples{
QRdecompose(rbind(c(2,-2,18),c(2,1,0),c(1,2,0)))
}
