% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SolveRationalMatrixEquaton.R
\name{sol.rationalmatrix.euqation}
\alias{sol.rationalmatrix.euqation}
\title{Solve Rational Matrix Equation}
\usage{
sol.rationalmatrix.euqation(Q, L, num_iterations = 50)
}
\arguments{
\item{Q}{a symmetrix postive definite matrix of real numbers}

\item{L}{a non-singular matrix of real numbers}

\item{num_iterations}{Number of iterations to run for convergence}
}
\value{
X : solution to the equation X = Q + L (X inv) L^T
}
\description{
Given a symmetrix positive definite matrix Q and a non-singular matrix L, Find symmetric positive definite solution X such that X = Q + L (X inv) L^T
}
\examples{
sol.rationalmatrix.euqation(matrix(c(2,-1,-1,2), 2, 2), rbind(c(2,3),c(2,1)))
}
